+++ guoba.support.js
@@ -1,9 +1,7 @@
-import { supportGuoba } from "./model/guoba/index.js"
-
 /**
  * 支持锅巴
  * 锅巴插件：https://gitee.com/guoba-yunzai/guoba-plugin.git
  * 组件类型，可参考 https://antdv.com/components/overview-cn/
  */
 
-export { supportGuoba }
+export { supportGuoba } from "#model"